ui beta version added

+++ index.js
@@ -19,17 +19,13 @@ app.set("view engine", "ejs")
 
 // main page get request handler
 router.get('/', function (req, res) {
-  upstreams.map(async (upstream, i) => {
-    if (upstream.mode === 'tcp')
-      await collectTCPInfo(upstream, i)
-    else
-      upstream.res = `${i + 1}. any mode except for TCP (${upstream.mode}) not currently supported!`
 
+  collectTCPInfo().finally(() => {
+    res.render("index", {
+      data: upstreams,
+    }); // index refers to index.ejs
   })
 
-  res.render("index", {
-    data: upstreams,
-  }); // index refers to index.ejs
 });
 
 //add the / router
@@ -40,28 +36,46 @@ app.listen(process.env.port || 3000);
 console.log('Running at Port 3000');
 
 // function to make tcp connection to an upstream
-const collectTCPInfo = (upstream, i) => {
-  return new Promise((resolve) => {
+const collectTCPInfo = async () => {
+  const promises = []
 
-    const client = new net.Socket();
+  upstreams.map(async (upstream, i) => {
+    if (upstream.mode === 'tcp') {
+      promises[i] = new Promise((resolve) => {
 
-    client.setTimeout(TIMEOUT, () => {
-      // Logs when could not make connection within X secs
-      upstream.res = `${i + 1}. ${upstream.name}:\t Connection timedout \t ${upstream.host}:${upstream.port}`
-      client.destroy()
-      resolve('timedout')
-    })
+        const client = new net.Socket();
 
-    client.connect(upstream.port, upstream.host, () => {
-      // Logs when the connection is established
-      upstream.res = `${i + 1}. ${upstream.name}:\t Connected \t\t ${upstream.host}:${upstream.port}`
-      client.destroy()
-      resolve(upstream)
-    });
+        client.setTimeout(TIMEOUT, () => {
+          // when could not make connection within X secs
+          upstream.resultText = `Connection timedout`
+          upstream.result = false
+          client.destroy()
+          resolve('timedout')
+        })
 
-    client.on("error", () => {
-      upstream.res = 'inaccessible'
-    })
+        client.connect(upstream.port, upstream.host, () => {
+          // when the connection is established
+          upstream.resultText = `Connected`
+          upstream.result = true
+          client.destroy()
+          resolve(upstream)
+        });
+
+        client.on("error", () => {
+          // when there is an error making the connection
+          upstream.resultText = 'inaccessible'
+          upstream.result = false
+          client.destroy()
+          resolve(upstream)
+        })
+      })
+    } else {
+      upstream.resultText = `any mode except for TCP (${upstream.mode}) not currently supported!`
+      upstream.result = false
+    }
+
   })
+
+  return await Promise.all(promises)
 
 }
